refactor(storage): extract buffer pool into bufferPool type
Move the pooled inputBuf and its get/free into buffer_pool.go behind a
bufferPool type with a Get method, so a shared pool is a one-line
newBufferPool init and the buffer owns the put back to its pool.
